Moved cv stuff out of cvwrapper to util/cv
