+++ opsi/modules/draw/fps.py
@@ -5,7 +5,7 @@ import cv2
 import numpy as np
 
 from opsi.manager.manager_schema import Function
-from opsi.manager.types import Mat
+from opsi.util.cv import Mat
 
 
 class FPS:
