+++ opsi/modules/nt.py
@@ -2,7 +2,8 @@ from dataclasses import dataclass
 
 from opsi.manager.manager_schema import Function
 from opsi.manager.netdict import NetworkDict
-from opsi.manager.types import AnyType, Point
+from opsi.manager.types import AnyType
+from opsi.util.cv import Point
 from opsi.util.unduplicator import Unduplicator
 
 __package__ = "opsi.nt"
