+++ opsi/modules/videoio/__init__.py
@@ -1,7 +1,6 @@
 from dataclasses import dataclass
 
 from opsi.manager.manager_schema import Function
-from opsi.manager.types import Mat
 from opsi.util.unduplicator import Unduplicator
 
 from .cameraserver import CameraSource, CamHook
@@ -9,6 +8,8 @@ from .input import controls, create_capture, get_modes, parse_camstring
 
 __package__ = "opsi.videoio"
 __version__ = "0.123"
+
+from opsi.util.cv import Mat
 
 UndupeInstance = Unduplicator()
 HookInstance = CamHook()
